Remove azure functions support, focusing only on AWS in this project.
