--- mangum/__main__.py
@@ -1,8 +1,0 @@
-def main() -> None:  # pragma: no cover
-    from mangum.cli import mangum
-
-    mangum()
-
-
-if __name__ == "__main__":  # pragma: no cover
-    main()
