+++ mangum/__version__.py
@@ -1,1 +1,1 @@
-__version__ = "0.6.4"  # pragma: no cover
+__version__ = "0.6.5"  # pragma: no cover
