+++ requirements.txt
@@ -1,6 +1,3 @@
-# Optional
-azure-functions
-
 # Testing
 pytest
 codecov
